# docs/api_documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DebtBot AI API",
    "version": "1.0.0",
    "description": "AI-powered multilingual debt collection chatbot for India"
  },
  "servers": [
    {
      "url": "https://your-domain.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/whatsapp/webhook": {
      "get": {
        "summary": "Verify WhatsApp webhook",
        "description": "Webhook verification endpoint for WhatsApp",
        "parameters": [
          { "name": "hub.mode", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "hub.verify_token", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "hub.challenge", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Webhook verified" },
          "403": { "description": "Verification failed" }
        }
      },
      "post": {
        "summary": "Handle WhatsApp webhook",
        "description": "Process incoming WhatsApp messages and status updates",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "WhatsApp webhook payload"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook processed successfully" }
        }
      }
    }
  },
...
}